+++ subprojects/api/src/main/groovy/org/pipelinelabs/pipeline/api/Stage.java
@@ -1,4 +1,4 @@
-package org.pipelinelabs.pipeline.runner.api;
+package org.pipelinelabs.pipeline.api;
 
 public interface Stage {
     void add(Task task);
