+++ subprojects/api/src/main/groovy/org/pipelinelabs/pipeline/api/Task.java
@@ -1,0 +1,4 @@
+package org.pipelinelabs.pipeline.api;
+
+public interface Task {
+}
