+++ subprojects/api/src/main/groovy/org/pipelinelabs/pipeline/api/task/ShellCommand.java
@@ -1,0 +1,6 @@
+package org.pipelinelabs.pipeline.api.task;
+
+import org.pipelinelabs.pipeline.api.Task;
+
+public interface ShellCommand extends Task {
+}
