+++ subprojects/runner/runner.gradle
@@ -9,6 +9,8 @@ mainClassName = 'org.pipelinelabs.pipeline.runner.cli.Pipeline'
 dependencies {
     compile libraries.groovy
     compile project(':api')
+    compile project(':groovydsl')
+    runtime project(':configmodel')
 
     compile libraries.jcommander
     compile libraries.mavenSharedUtils
