--- subprojects/runner/src/main/groovy/org/pipelinelabs/pipeline/runner/api/Task.java
@@ -1,4 +1,0 @@
-package org.pipelinelabs.pipeline.runner.api;
-
-public interface Task {
-}
