--- subprojects/runner/src/main/groovy/org/pipelinelabs/pipeline/runner/api/task/ShellCommand.java
@@ -1,6 +1,0 @@
-package org.pipelinelabs.pipeline.runner.api.task;
-
-import org.pipelinelabs.pipeline.runner.api.Task;
-
-public interface ShellCommand extends Task {
-}
