+++ subprojects/runner/src/main/groovy/org/pipelinelabs/pipeline/runner/dsl/internal/DefaultStageDsl.java
@@ -4,7 +4,7 @@ import org.apache.maven.shared.utils.cli.CommandLineException;
 import org.apache.maven.shared.utils.cli.CommandLineUtils;
 import org.apache.maven.shared.utils.cli.Commandline;
 import org.apache.maven.shared.utils.cli.WriterStreamConsumer;
-import org.pipelinelabs.pipeline.runner.api.Stage;
+import org.pipelinelabs.pipeline.api.Stage;
 import org.pipelinelabs.pipeline.runner.runner.DefaultStage;
 
 import java.io.BufferedWriter;
