+++ subprojects/runner/src/main/groovy/org/pipelinelabs/pipeline/runner/dsl/internal/InternalPipelineDsl.java
@@ -1,6 +1,6 @@
 package org.pipelinelabs.pipeline.runner.dsl.internal;
 
-import org.pipelinelabs.pipeline.runner.api.Pipeline;
+import org.pipelinelabs.pipeline.api.Pipeline;
 import org.pipelinelabs.pipeline.runner.dsl.PipelineDsl;
 
 public interface InternalPipelineDsl extends PipelineDsl, DslExporter<Pipeline> {
