+++ subprojects/runner/src/main/groovy/org/pipelinelabs/pipeline/runner/dsl/internal/InternalStageDsl.java
@@ -1,6 +1,6 @@
 package org.pipelinelabs.pipeline.runner.dsl.internal;
 
-import org.pipelinelabs.pipeline.runner.api.Stage;
+import org.pipelinelabs.pipeline.api.Stage;
 import org.pipelinelabs.pipeline.runner.dsl.StageDsl;
 
 public interface InternalStageDsl extends StageDsl, DslExporter<Stage> {
