+++ subprojects/runner/src/main/groovy/org/pipelinelabs/pipeline/runner/dsl/internal/InternalTaskDsl.java
@@ -1,6 +1,6 @@
 package org.pipelinelabs.pipeline.runner.dsl.internal;
 
-import org.pipelinelabs.pipeline.runner.api.Task;
+import org.pipelinelabs.pipeline.api.Task;
 import org.pipelinelabs.pipeline.runner.dsl.TaskDsl;
 
 public interface InternalTaskDsl extends TaskDsl, DslExporter<Task> {
