+++ subprojects/runner/src/main/groovy/org/pipelinelabs/pipeline/runner/runner/DefaultStage.java
@@ -1,7 +1,7 @@
 package org.pipelinelabs.pipeline.runner.runner;
 
-import org.pipelinelabs.pipeline.runner.api.Stage;
-import org.pipelinelabs.pipeline.runner.api.Task;
+import org.pipelinelabs.pipeline.api.Stage;
+import org.pipelinelabs.pipeline.api.Task;
 
 public class DefaultStage implements Stage {
